project finished

+++ libftprintf/ft_puthxd.c
@@ -6,14 +6,14 @@
 /*   By: darguerr <darguerr@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/30 17:53:39 by darguerr          #+#    #+#             */
-/*   Updated: 2025/06/03 13:53:19 by darguerr         ###   ########.fr       */
+/*   Updated: 2025/06/04 23:55:23 by darguerr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
 /* This function print a decimal number in hexadecimal number */
-int	ft_puthxd(unsigned int num, int uppercase)
+int	ft_puthxd(unsigned long num, int uppercase)
 {
 	const char	*hex;
 	int			count;
